stress-sigfd.c: Fix compile warning
cc -Wall -Wextra -DVERSION="0.04.12" -O2 -c -o stress-sigfd.o stress-sigfd.c
stress-sigfd.c: In function ‘stress_sigfd’:
stress-sigfd.c:110:24: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     if (fdsi.ssi_signo != SIGRTMIN) {
                        ^

+++ stress-sigfd.c
@@ -107,7 +107,7 @@ again:
 			if (ret == 0)
 				break;
 			if (opt_flags & OPT_FLAGS_VERIFY) {
-				if (fdsi.ssi_signo != SIGRTMIN) {
+				if (fdsi.ssi_signo != (uint32_t)SIGRTMIN) {
 					pr_fail(stderr, "%s: unexpected signal %d",
 						name, fdsi.ssi_signo);
 					break;
